requirements.txt
asgiref==3.2.7
Babel==2.8.0
Django==3.0.7
django-crispy-forms==1.9.1
django-phonenumber-field==4.0.0
phonenumberslite==8.12.5
pytz==2020.1
sqlparse==0.3.1
